+++ Opcodes/babo.c
@@ -159,7 +159,7 @@ BaboMemory_create(CSOUND *csound, BaboMemory *this, size_t size_in_floats)
 {
     size_t size_in_bytes = size_in_floats * sizeof(MYFLT);
 
-    csound->AuxAlloc(csound, (long) size_in_bytes, &this->memptr);
+    csound->AuxAlloc(csound, size_in_bytes, &this->memptr);
 
     memset(this->memptr.auxp, 0, size_in_bytes);
